Some code restructure (#301)
* Moved modules a bit

* Factor http functionality into a package

* fix linter warnings in ota package

* Change default config to point to a new version
