--- ota/version.go
@@ -1,6 +1,0 @@
-package ota
-
-var (
-	Version    = "v0.0.1"
-	Repository = "Arriven/db1000n" // Could be changed via the ldflags
-)
