+++ src/core/dnsblast/qry/types.go
