+++ src/utils/ota/version.go
@@ -1,0 +1,8 @@
+package ota
+
+var (
+	// Version is a releae version embedded into the app
+	Version = "v0.0.1"
+	// Repository to check for updates
+	Repository = "Arriven/db1000n" // Could be changed via the ldflags
+)
